program works now.  May fail to find all roots for e.g. p=2,3 but we dont care.  Program is aimed at bigger primes (we do not sieve the smallest primes in the factor base of the number field sieve).

+++ intpoly.h
@@ -8,7 +8,7 @@ using std::vector;
 int primitiveroot(int p, int* q);
 int modpow(int64_t r, int64_t e, int64_t p);
 int chienrootsmod(int* f, int degf, int p, int* r, int* q);
-int mod(int r, int p);
+int mod(int64_t r, int64_t p);
 inline int modinv(int x, int m);
 inline int poldegree(int64_t* f, int maxd);
 inline void polsetzero(int64_t* f, int maxd);
